+++ build-packages/magento-scripts/lib/tasks/php-fpm/stop-php-fpm.js
@@ -1,6 +1,8 @@
 const fs = require('fs');
 const { execAsyncSpawn } = require('../../util/exec-async-command');
 const pathExists = require('../../util/path-exists');
+const getPhpConfig = require('../../config/php');
+const { getBaseConfig } = require('../../config/index');
 
 const getProcessId = async (fpmPidFilePath) => {
     const pidExists = await pathExists(fpmPidFilePath);
@@ -17,22 +19,15 @@ const getProcessId = async (fpmPidFilePath) => {
  */
 const stopPhpFpmTask = {
     title: 'Stopping php-fpm',
-    task: async ({ config: { php } }, task) => {
+    task: async ({ config: { overridenConfiguration }, projectPath }, task) => {
+        const php = getPhpConfig(overridenConfiguration.configuration, getBaseConfig(projectPath));
+        const processId = await getProcessId(php.fpmPidFilePath);
+        if (!processId) {
+            task.skip();
+            return;
+        }
         try {
-            const processId = await getProcessId(php.fpmPidFilePath);
-            if (!processId) {
-                task.skip();
-                return;
-            }
             await execAsyncSpawn(`kill ${processId}`);
-
-            if (await pathExists(php.fpmPidFilePath)) {
-                try {
-                    await fs.promises.unlink(php.fpmPidFilePath);
-                } catch (e) {
-                    //
-                }
-            }
         } catch (e) {
             if (e.toLowerCase().includes('no such process')) {
                 try {
@@ -43,11 +38,14 @@ const stopPhpFpmTask = {
 
                 return;
             }
+        }
 
-            throw new Error(
-                `Unexpected error while stopping php-fpm.
-                See ERROR log below.\n\n${e}`
-            );
+        if (await pathExists(php.fpmPidFilePath)) {
+            try {
+                await fs.promises.unlink(php.fpmPidFilePath);
+            } catch (e) {
+                //
+            }
         }
     },
     options: {