+++ build-packages/magento-scripts/lib/util/path-exists.js
@@ -1,5 +1,10 @@
 const fs = require('fs');
 
+/**
+ * Check if provided path exists in filesystem
+ * @param {string} path
+ * @returns {Promise<boolean>}
+ */
 const pathExists = async (path) => {
     try {
         await fs.promises.access(path, fs.constants.F_OK);