+++ build-packages/magento-scripts/lib/util/resolve-configuration-with-overrides.js
@@ -1,23 +1,30 @@
 const path = require('path');
 const { configFileSchema } = require('./config-file-validator');
 const deepmerge = require('./deepmerge');
+const pathExists = require('./path-exists');
 
-const configJSFilePath = path.join(process.cwd(), 'cma.js');
+const resolveConfigurationWithOverrides = async (configuration, baseConfig, projectPath = process.cwd()) => {
+    const configJSFilePath = path.join(projectPath, 'cma.js');
+    if (await pathExists(configJSFilePath)) {
+        const userConfiguration = require(configJSFilePath);
 
-const resolveConfigurationWithOverrides = async (configuration) => {
-    const userConfiguration = require(configJSFilePath);
+        try {
+            await configFileSchema.validateAsync(userConfiguration);
+        } catch (e) {
+            throw new Error(`Configuration file validation error!\n\n${e.message}`);
+        }
 
-    try {
-        await configFileSchema.validateAsync(userConfiguration);
-    } catch (e) {
-        throw new Error(`Configuration file validation error!\n\n${e.message}`);
+        const overridenConfiguration = deepmerge(configuration, userConfiguration);
+
+        return {
+            userConfiguration,
+            overridenConfiguration
+        };
     }
 
-    const overridenConfiguration = deepmerge(configuration, userConfiguration);
-
     return {
-        userConfiguration,
-        overridenConfiguration
+        userConfiguration: configuration,
+        overridenConfiguration: configuration
     };
 };